fix various bugs
