+++ packages/core-mobile/app/new/common/components/HiddenBalanceText.tsx
@@ -1,24 +1,42 @@
 import { SxProp, Text, TextVariant } from '@avalabs/k2-alpine'
 import { useFormatCurrency } from 'common/hooks/useFormatCurrency'
-import React from 'react'
+import React, { useMemo } from 'react'
 
 export const HiddenBalanceText = ({
   variant = 'heading2',
+  isCurrency = true,
   sx
 }: {
   variant?: TextVariant
+  isCurrency?: boolean
   sx?: SxProp
 }): React.JSX.Element => {
   const { formatCurrency } = useFormatCurrency()
-  const maskedText = formatCurrency({
-    amount: 0
-  })
-    .replace('0.00', '')
-    .trim()
+  const maskedText = useMemo(() => {
+    const dots = dot.repeat(6)
+
+    if (isCurrency) {
+      return formatCurrency({ amount: 0 }).replace(/[\d.,]+/g, dots)
+    }
+
+    return dots
+  }, [formatCurrency, isCurrency])
 
   return (
-    <Text variant={variant} sx={sx}>
-      {maskedText + '••••••'}
+    <Text>
+      {Array.from(maskedText).map((char, i) => (
+        <Text
+          key={i}
+          variant={variant}
+          sx={{
+            ...(char === dot ? { fontFamily: 'DejaVuSansMono' } : {}),
+            ...sx
+          }}>
+          {char}
+        </Text>
+      ))}
     </Text>
   )
 }
+
+const dot = '•'
