+++ packages/core-mobile/app/new/common/components/SubTextNumber.tsx
@@ -1,18 +1,24 @@
 import React, { useMemo } from 'react'
 import { View, Text, alpha, useTheme } from '@avalabs/k2-alpine'
-import { StyleSheet } from 'react-native'
+import { StyleSheet, ViewStyle } from 'react-native'
 import { numberToSubscriptFormat } from 'utils/numberToSubscriptFormat/numberToSubscriptFormat'
 
 interface Props {
   number: number | undefined
   testID?: string
-  size?: 'big' | 'small'
+  textColor?: string
+  style?: ViewStyle
+  textSize?: number | undefined
+  subTextSize?: number | undefined
 }
 
 export const SubTextNumber: React.FC<Props> = ({
   number,
   testID,
-  size = 'small'
+  textColor,
+  style,
+  textSize = 16,
+  subTextSize = 13
 }) => {
   const {
     theme: { colors }
@@ -22,30 +28,26 @@ export const SubTextNumber: React.FC<Props> = ({
     [number]
   )
 
-  const textVariant = size === 'big' ? 'heading5' : 'body1'
-  const textSize = size === 'big' ? 26 : 16
-  const textColor = alpha(colors.$textPrimary, 0.6)
+  const _textColor = textColor ?? alpha(colors.$textPrimary, 0.6)
 
   return (
-    <View style={styles.container} testID={testID}>
+    <View style={[styles.container, style]} testID={testID}>
       <Text
         numberOfLines={1}
-        variant={textVariant}
-        sx={{ color: textColor, fontSize: textSize }}>
+        style={[{ color: _textColor, fontSize: textSize }]}>
         {mainTextBefore}
       </Text>
       {subText && (
         <Text
-          style={size === 'big' ? styles.subTextBig : styles.subTextSmall}
-          variant={textVariant}
-          sx={{ color: textColor }}>
+          style={[
+            styles.subTextSmall,
+            { color: _textColor, fontSize: subTextSize }
+          ]}>
           {subText}
         </Text>
       )}
       {mainTextAfter && (
-        <Text
-          variant={textVariant}
-          sx={{ color: textColor, fontSize: textSize }}>
+        <Text style={[{ color: _textColor, fontSize: textSize }]}>
           {mainTextAfter}
         </Text>
       )}
@@ -57,16 +59,11 @@ const styles = StyleSheet.create({
   container: {
     flexDirection: 'row',
     alignItems: 'center',
-    maxWidth: '70%'
+    maxWidth: '70%',
+    overflow: 'visible'
   },
   subTextSmall: {
-    fontSize: 13,
     fontWeight: '500',
-    position: 'relative',
-    top: 4
-  },
-  subTextBig: {
-    fontSize: 15,
     position: 'relative',
     top: 4
   }
