+++ packages/core-mobile/app/new/routes/(signedIn)/(tabs)/portfolio/index.tsx
@@ -91,16 +91,14 @@ const PortfolioHomeScreen = (): JSX.Element => {
   )
   const selectedCurrency = useSelector(selectSelectedCurrency)
   const { formatCurrency } = useFormatCurrency()
-  const currencyBalance = useMemo(() => {
+  const formattedBalance = useMemo(() => {
     return !balanceAccurate || balanceTotalInCurrency === 0
-      ? '$' + UNKNOWN_AMOUNT
-      : formatCurrency({ amount: balanceTotalInCurrency })
+      ? UNKNOWN_AMOUNT
+      : formatCurrency({
+          amount: balanceTotalInCurrency,
+          withoutCurrencySuffix: true
+        })
   }, [balanceAccurate, balanceTotalInCurrency, formatCurrency])
-
-  const formattedBalance = useMemo(
-    () => currencyBalance.replace(selectedCurrency, ''),
-    [currencyBalance, selectedCurrency]
-  )
 
   const { getMarketTokenBySymbol } = useWatchlist()
   const tokens = useFocusedSelector((state: RootState) =>
