+++ packages/k2-alpine/src/components/index.ts
@@ -21,6 +21,7 @@ export * from './Dropdown/SimpleDropdown'
 export * from './PriceChangeIndicator/PriceChangeIndicator'
 export * from './Header/BalanceHeader'
 export * from './Header/NavigationTitleHeader'
+export * from './Header/PrivacyModeAlert'
 export * from './Dropdown/usePopoverAnchor'
 export * from './Toggle/Toggle'
 export * from './Animated/AnimatedText'
